Move component to CardWithAccordionComponent
